refactor(thermidor): remove #sourceEngine and enforce internal/public boundary
