+++ packages/thermidor/src/internal/interfaces/index.ts
@@ -1,0 +1,5 @@
+export {CommerceInterfaceImpl} from './commerce.js';
+export {SearchInterfaceImpl} from './search.js';
+export {GenerativeInterfaceImpl} from './generative.js';
+export {ComposedInterfaceImpl, getComposedInternals} from './compose.js';
+export type {ComposedInternals} from './compose.js';
