+++ packages/thermidor/src/public/interfaces/compose.test.ts
@@ -88,7 +88,9 @@ describe('composeInterfaces', () => {
       interfaces: [searchInterface, generativeInterface],
     });
 
-    const thunks = getComposedInternals(composed).resolveFacades('search');
+    const thunks = getComposedInternals(
+      composed as ComposedInterfaceImpl<'search' | 'generative'>
+    ).resolveFacades('search');
     expect(thunks).toHaveLength(1);
     expect(thunks[0]).toBe(mockSearchThunk);
   });
@@ -108,7 +110,9 @@ describe('composeInterfaces', () => {
     const ifaceA = new TestSearchInterface(engine, 'a');
 
     const composed = composeInterfaces({interfaces: [ifaceA]});
-    const {stateId} = getComposedInternals(composed);
+    const {stateId} = getComposedInternals(
+      composed as ComposedInterfaceImpl<'search'>
+    );
 
     expect(typeof stateId).toBe('string');
     expect(stateId.length).toBeGreaterThan(0);
