+++ packages/thermidor/src/public/interfaces/compose.ts
@@ -1,101 +1,18 @@
-import {InterfaceCacheRegistry} from '@/src/internal/utils/index.js';
 import {getHandleInternals} from '@/src/internal/utils/index.js';
-import type {FullEngine} from '@/src/internal/engine/index.js';
+import {generateId} from '@/src/internal/utils/index.js';
 import type {
-  EndpointThunk,
-  Facades,
+  ComposedInterface,
   InferInterfaceType,
-  InterfaceHandle,
   InterfaceType,
   InterfaceTypeMap,
-  Supports,
-  SupportsBrand,
 } from '@/src/internal/utils/index.js';
-import {generateId} from '@/src/internal/utils/index.js';
+import {ComposedInterfaceImpl} from '@/src/internal/interfaces/index.js';
 
-export interface ComposedInternals {
-  engine: FullEngine;
-  stateId: string;
-  cacheRegistry: InterfaceCacheRegistry;
-  resolveFacades(
-    facade: Facades[InterfaceType],
-    composedInterface?: InterfaceHandle
-  ): EndpointThunk[];
-}
-
-export let getComposedInternals: <T extends InterfaceType>(
-  composed: ComposedInterfaceImpl<T>
-) => ComposedInternals;
-
-export interface ComposedInterface<T extends InterfaceType> extends Supports<
-  Facades[T]
-> {}
-
-export class ComposedInterfaceImpl<
-  T extends InterfaceType,
-> implements ComposedInterface<T> {
-  declare readonly [SupportsBrand]: {[K in Facades[T]]: true};
-
-  get disposed(): boolean {
-    return this.#disposed;
-  }
-
-  #engine: FullEngine;
-  #stateId: string;
-  #interfaces: InterfaceHandle[];
-  #cacheRegistry = new InterfaceCacheRegistry();
-  #disposed = false;
-
-  static {
-    getComposedInternals = <typeof getComposedInternals>((composed) => ({
-      engine: composed.#engine,
-      stateId: composed.#stateId,
-      cacheRegistry: composed.#cacheRegistry,
-      resolveFacades: (facade: any, composedInterface?: InterfaceHandle) =>
-        composed.#resolveFacades(facade, composedInterface),
-    }));
-  }
-
-  constructor(interfaces: InterfaceHandle[], composedId: string) {
-    if (interfaces.length === 0) {
-      throw new Error('ComposedInterface requires at least one interface.');
-    }
-
-    const {engine} = getHandleInternals(interfaces[0]);
-    this.#engine = engine;
-    this.#stateId = composedId;
-    this.#interfaces = interfaces;
-
-    engine.addInterface(this);
-  }
-
-  dispose(): void {
-    if (this.#disposed) {
-      return;
-    }
-    this.#disposed = true;
-    this.#cacheRegistry.dispose();
-    this.#engine.removeInterface(this);
-  }
-
-  #resolveFacades(
-    facade: Facades[T],
-    composedInterface?: InterfaceHandle
-  ): EndpointThunk[] {
-    this.#assertNotDisposed();
-    const target = composedInterface ?? this;
-    return this.#interfaces.flatMap((sub) => {
-      const {resolveFacades} = getHandleInternals(sub);
-      return resolveFacades(facade, target);
-    });
-  }
-
-  #assertNotDisposed(): void {
-    if (this.#disposed) {
-      throw new Error('Cannot operate on a disposed composed interface.');
-    }
-  }
-}
+export type {ComposedInterface} from '@/src/internal/utils/index.js';
+export {
+  ComposedInterfaceImpl,
+  getComposedInternals,
+} from '@/src/internal/interfaces/index.js';
 
 export function composeInterfaces<
   I extends InterfaceTypeMap[InterfaceType],
