+++ packages/thermidor/src/public/interfaces/search.ts
@@ -1,31 +1,9 @@
-import {BaseInterface} from '@/src/internal/utils/index.js';
-import type {FullEngine} from '@/src/internal/engine/index.js';
 import {Engine, getFullEngine} from '@/src/internal/engine/index.js';
-import type {
-  FacadeResolverFactory,
-  Facades,
-  Supports,
-} from '@/src/internal/utils/index.js';
 import {generateId} from '@/src/internal/utils/index.js';
-import {getOrCreateSearchParametersSlice} from '@/src/internal/features/search-parameters/index.js';
-import {createSearchFacadeResolver} from '@/src/internal/api/search/index.js';
-import {createQuerySuggestFacadeResolver} from '@/src/internal/api/query-suggest/index.js';
+import type {SearchInterface} from '@/src/internal/utils/index.js';
+import {SearchInterfaceImpl} from '@/src/internal/interfaces/index.js';
 
-const resolverFactories: Record<Facades['search'], FacadeResolverFactory> = {
-  search: createSearchFacadeResolver,
-  suggestions: createQuerySuggestFacadeResolver,
-};
-
-export interface SearchInterface extends Supports<Facades['search']> {}
-
-export class SearchInterfaceImpl
-  extends BaseInterface<'search'>
-  implements SearchInterface
-{
-  constructor(engine: FullEngine, stateId: string) {
-    super(engine, stateId, 'search', resolverFactories);
-  }
-}
+export type {SearchInterface} from '@/src/internal/utils/index.js';
 
 export interface BuildSearchInterfaceOptions {
   engine: Engine;
@@ -38,9 +16,5 @@ export function buildSearchInterface(
   const fullEngine = getFullEngine(options.engine);
   const interfaceId = options.id ?? generateId();
 
-  const iface = new SearchInterfaceImpl(fullEngine, interfaceId);
-
-  fullEngine.adoptSlice(getOrCreateSearchParametersSlice(iface));
-
-  return iface;
+  return new SearchInterfaceImpl(fullEngine, interfaceId);
 }
